MindMap/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
bert-score==0.3.13
Bottleneck 
brotlipy==0.7.0
certifi==2023.7.22
cffi
chardet==5.1.0
charset-normalizer==2.1.0
click==8.1.3
cmake==3.27.0
colorama
contourpy==1.0.7
cryptography
cycler==0.11.0
Cython==0.29.35
dataclasses
dataclasses-json==0.5.7
datasets==2.13.1
dill==0.3.6
evaluate==0.4.0
filelock 
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.6.0
gensim==4.3.1
greenlet==2.0.2
huggingface-hub==0.16.4
idna 
importlib-metadata 
importlib-resources==5.12.0
Jinja2==3.1.2
joblib 
kiwisolver==1.4.4
langchain==0.0.181
lit==16.0.6
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-random 
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess
mypy-extensions==1.0.0
neo4j-driver==5.9.0
networkx==3.1
nltk==3.8.1
numexpr
numpy 
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging
pandas==1.5.3
Pillow==9.5.0
pooch
protobuf==3.20.3
pyahocorasick==2.0.0
pyarrow==12.0.1
pycocoevalcap
pycocotools==2.0.6
pycparser
pydantic==1.10.8
pyOpenSSL 
pyparsing==3.0.9
PySocks
python-dateutil
pytz
PyYAML
rank-bm25==0.2.2
regex
requests
responses==0.18.0
sacremoses==0.0.43
scikit-learn
scipy==1.10.1
six
smart-open
SQLAlchemy==2.0.15
sympy==1.12
tenacity==8.2.2
threadpoolctl
tokenizers==0.13.0
torch==2.0.1
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3 
xxhash 
yarl==1.9.2
zipp==3.15.0
